requirements.txt
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.56.1
django-cors-headers==4.2.0
django-import-export==3.2.0
django-phonenumber-field==7.3.0
django-resized==1.0.2
django-rest-auth==0.9.5
whitenoise==6.6.0
phonenumbers==8.13.31
gunicorn==21.2.0
jsonschema==4.21.1
pandas==2.2.1
openpyxl==3.1.2